Setting up /app/v1/project-listing

+++ data/src/main/java/com/proptiger/data/model/image/Image.java
@@ -16,14 +16,14 @@ public class Image implements Serializable {
 	@Id
 	@Column(name = "id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
+	private int id;
 	
 	@ManyToOne
 	@JoinColumn(name = "ImageType_id")
 	private ImageType imageType;
 	
 	@Column(name = "object_id")
-	private long objectId;
+	private int objectId;
 	
 	@Column(name = "path")
 	private String path;
@@ -55,63 +55,39 @@ public class Image implements Serializable {
 	@Column(name = "seo_name")
 	private String seoName;
 
-	/**
-	 * @return the id
-	 */
-	public long getId() {
-		return id;
-	}
+    public int getId() {
+        return id;
+    }
 
-	/**
-	 * @param id the id to set
-	 */
-	public void setId(long id) {
-		this.id = id;
-	}
+    public void setId(int id) {
+        this.id = id;
+    }
 
-	/**
-	 * @return the imageType
-	 */
-	public ImageType getImageType() {
-		return imageType;
-	}
+    public ImageType getImageType() {
+        return imageType;
+    }
 
-	/**
-	 * @param imageType the imageType to set
-	 */
-	public void setImageType(ImageType imageType) {
-		this.imageType = imageType;
-	}
+    public void setImageType(ImageType imageType) {
+        this.imageType = imageType;
+    }
 
-	/**
-	 * @return the objectId
-	 */
-	public long getObjectId() {
-		return objectId;
-	}
+    public int getObjectId() {
+        return objectId;
+    }
 
-	/**
-	 * @param objectId the objectId to set
-	 */
-	public void setObjectId(long objectId) {
-		this.objectId = objectId;
-	}
+    public void setObjectId(int objectId) {
+        this.objectId = objectId;
+    }
 
-	/**
-	 * @return the path
-	 */
-	public String getPath() {
-		return path;
-	}
+    public String getPath() {
+        return path;
+    }
 
-	/**
-	 * @param path the path to set
-	 */
-	public void setPath(String path) {
-		this.path = path;
-	}
+    public void setPath(String path) {
+        this.path = path;
+    }
 
-	/**
+    /**
 	 * @return the createdAt
 	 */
 	public Date getCreatedAt() {
@@ -139,102 +115,60 @@ public class Image implements Serializable {
 		this.takenAt = takenAt;
 	}
 
-	/**
-	 * @return the size
-	 */
-	public long getSize() {
-		return size;
-	}
+    public long getSize() {
+        return size;
+    }
 
-	/**
-	 * @param size the size to set
-	 */
-	public void setSize(long size) {
-		this.size = size;
-	}
+    public void setSize(long size) {
+        this.size = size;
+    }
 
-	/**
-	 * @return the width
-	 */
-	public int getWidth() {
-		return width;
-	}
+    public int getWidth() {
+        return width;
+    }
 
-	/**
-	 * @param width the width to set
-	 */
-	public void setWidth(int width) {
-		this.width = width;
-	}
+    public void setWidth(int width) {
+        this.width = width;
+    }
 
-	/**
-	 * @return the height
-	 */
-	public int getHeight() {
-		return height;
-	}
+    public int getHeight() {
+        return height;
+    }
 
-	/**
-	 * @param height the height to set
-	 */
-	public void setHeight(int height) {
-		this.height = height;
-	}
+    public void setHeight(int height) {
+        this.height = height;
+    }
 
-	/**
-	 * @return the format
-	 */
-	public String getFormat() {
-		return format;
-	}
+    public String getFormat() {
+        return format;
+    }
 
-	/**
-	 * @param format the format to set
-	 */
-	public void setFormat(String format) {
-		this.format = format;
-	}
+    public void setFormat(String format) {
+        this.format = format;
+    }
 
-	/**
-	 * @return the order
-	 */
-	public Integer getOrder() {
-		return order;
-	}
+    public Integer getOrder() {
+        return order;
+    }
 
-	/**
-	 * @param order the order to set
-	 */
-	public void setOrder(Integer order) {
-		this.order = order;
-	}
+    public void setOrder(Integer order) {
+        this.order = order;
+    }
 
-	/**
-	 * @return the contentHash
-	 */
-	public String getContentHash() {
-		return contentHash;
-	}
+    public String getContentHash() {
+        return contentHash;
+    }
 
-	/**
-	 * @param contentHash the contentHash to set
-	 */
-	public void setContentHash(String contentHash) {
-		this.contentHash = contentHash;
-	}
+    public void setContentHash(String contentHash) {
+        this.contentHash = contentHash;
+    }
 
-	/**
-	 * @return the seoName
-	 */
-	public String getSeoName() {
-		return seoName;
-	}
+    public String getSeoName() {
+        return seoName;
+    }
 
-	/**
-	 * @param seoName the seoName to set
-	 */
-	public void setSeoName(String seoName) {
-		this.seoName = seoName;
-	}
+    public void setSeoName(String seoName) {
+        this.seoName = seoName;
+    }
 
 }
